somehash/proof-of-concept/requirements.txt
asgiref==3.8.1
contourpy==1.3.1
cycler==0.12.1
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
fonttools==4.55.3
joblib==1.4.2
kiwisolver==1.4.8
Markdown==3.7
matplotlib==3.10.0
numpy==2.2.1
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
pillow-avif-plugin==1.4.6
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pywatchman==2.0.0
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.5.0
wheel @ file:///opt/homebrew/Cellar/python%403.13/3.13.0_1/libexec/wheel-0.44.0-py3-none-any.whl#sha256=f49b82715dce6365f75eddcb0a9bb47d0d46feaf14bbc739dfcbd677b7073f5b
